+++ satstac/catalog.py
@@ -15,12 +15,12 @@ class Catalog(Thing):
     @property
     def stac_version(self):
         """ Get the STAC version of this catalog """
-        return self.data['stac_version']
+        return self._data['stac_version']
 
     @property
     def description(self):
         """ Get catalog description """
-        return self.data.get('description', '')
+        return self._data.get('description', '')
 
     @classmethod
     def create(cls, id='stac-catalog', title='A STAC Catalog', 
@@ -52,7 +52,7 @@ class Catalog(Thing):
     def collections(self):
         """ Recursively get all collections within this Catalog """
         for cat in self.children():
-            if 'extent' in cat.data.keys():
+            if 'extent' in cat._data.keys():
                 yield Collection.open(cat.filename)
                 # TODO - keep going? if other Collections can appear below a Collection
             else:
@@ -73,39 +73,18 @@ class Catalog(Thing):
         child_link = '%s/catalog.json' % catalog.id
         child_fname = os.path.join(self.path, child_link)
         child_path = os.path.dirname(child_fname)
-        root_link = self.links('root')[0]
+        root_links = self.links('root')
+        root_link = root_links[0] if len(root_links) > 0 else self.filename
         root_path = os.path.dirname(root_link)
         self.add_link('child', child_link)
         self.save()
         # strip self, parent, child links from catalog and add new links
         catalog.clean_hierarchy()
-        catalog.add_link('self', os.path.join(self.endpoint(), os.path.relpath(child_fname, root_path)))
         catalog.add_link('root', os.path.relpath(root_link, child_path))
         catalog.add_link('parent', os.path.relpath(self.filename, child_path))
         # create catalog file
-        catalog.save_as(child_fname)
+        catalog.save(filename=child_fname)
         return self
-
-    def endpoint(self):
-        """ Get endpoint URL to the root catalog """
-        return os.path.dirname(self.root().links('self')[0])
-
-    def publish(self, endpoint, root=None):
-        """ Update all self links throughout catalog to use new endpoint """
-        # we don't use the catalogs and items functions as we'd have to go 
-        # through the tree twice, once for catalogs and once for items
-        # update myself
-        if root is None:
-            root = self.filename
-        super(Catalog, self).publish(endpoint, root=root)
-        # update direct items
-        for link in self.links('item'):
-            item = Item.open(link)
-            item.publish(endpoint, root=root)
-        # follow children
-        for cat in self.children():
-            cat.publish(endpoint, root=root)
-
 
 
 # import and end of module prevents problems with circular dependencies.